Bug fixes for linux

+++ src/AveragePoolFilter.cpp
@@ -41,7 +41,7 @@ namespace tgr {
 	void AveragePoolFilter::initialize(NeuralSystem& sys, const NeuronFunction& func) {
 		std::vector<SignalPtr> signals;
 		outputLayers.resize(inputLayers.size());
-		for (int k = 0; k < inputLayers.size(); k++) {
+		for (int k = 0; k < (int)inputLayers.size(); k++) {
 			NeuralLayerPtr inputLayer = inputLayers[k];
 			outputLayers[k] = NeuralLayerPtr(new NeuralLayer(name,inputLayer->width/kernelSize,inputLayer->height/kernelSize,1,bias, func));
 			NeuralLayerPtr outputLayer = outputLayers[k];
